We work for understanding the factors and the business in detail and predict the price of used vehicles based on historical data provided
Used car market is huge and growing more than ever in India. Anyone who has bought or sold a used car will agree that one of the biggest challenge in the trade is estimating a fair and market smart price which would benefit both buyer and the seller.  Consequently, other than individuals, automobile manufactures, used cars dealers and loan lending banks would also be interested in a model which would help them predict right price of the car
Other than regular buy and sell, car lease business model is also very prevalent and helps individual to obtain a car by paying monthly “rent” on the car instead of buying it.  Considering the complex dynamics of used car sales business, a machine learning model could be appropriate to estimate the price of a car.